Upd TaskTwo replace unfold char methods to utility package
